feat: Add offline friend argument and refactor package structure for friend commands
