+++ surf-friends-paper/src/main/kotlin/dev/slne/surf/friends/paper/command/FriendCommand.kt
@@ -3,6 +3,7 @@ package dev.slne.surf.friends.paper.command
 import dev.jorel.commandapi.kotlindsl.commandTree
 import dev.jorel.commandapi.kotlindsl.literalArgument
 import dev.slne.surf.core.api.paper.command.argument.surfOfflinePlayerArgument
+import dev.slne.surf.friends.paper.command.argument.friend.offlineFriendArgument
 import dev.slne.surf.friends.paper.permission.PermissionRegistry
 
 fun friendCommand() = commandTree("friend") {
@@ -15,7 +16,9 @@ fun friendCommand() = commandTree("friend") {
     }
 
     literalArgument("remove") {
+        offlineFriendArgument("friend") {
 
+        }
     }
 
     literalArgument("list") {
